linkAlias change:
- new article field linkAlias "Link alias", this field must contain
only alphanumerics and at least one letter

- articles with a linkAlias have a url generated for them based on the
alias and their title.  You will need to use an ErrorDocument or
RewriteRule to point these at page.pl

- page.pl now accepts an alias parameter, and will display the article
with that alias.

- page.pl can now act as an Apache ErrorDocument handler and if the
original URI was a linkAlias url it will extract the alias and display
the given article.  Unfortunately Apache will still write an error to
the error log.