fix some problems with dyntarget
also make tag formats required to be at the end of the tag, so:

  <:tag foo |x :>

no longer works, but:

  <:tag foo |x:>

does.

This allows constructs like:

  <:tag [foo bar |x] :>

to be parsed so the x format is applied to the results of <:foo bar:>.

page.pl was calling BSE::UI::Page::dispatch as a class method, but it
makes object method calls now.